% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-helpers.R
\name{acs.demographic.recode}
\alias{acs.demographic.recode}
\title{acs.demographic.recode}
\usage{
acs.demographic.recode(demos, other.vars = c(5, 7:11), filter.aggregates = T)
}
\arguments{
\item{demos}{a demographics table, as from \code{multiyr.acs.wrapper(B03002, ...)}.}

\item{other.vars}{Vars (number following underscore) for groups to put into an "Other group" category.}

\item{filter.aggregates}{If true, filter totals and subtotals.}
}
\description{
Recodes demographic info, as from table B03002. Was developed for 2019 ACS;
could break if they change encodings.
}
\details{
Universe is Total Population.

Encodings this was developed for:
B03002_001	Total
B03002_002	Total: Not Hispanic or Latino
B03002_003	Total: Not Hispanic or Latino: White alone
B03002_004	Total: Not Hispanic or Latino: Black or African American alone
B03002_005	Total: Not Hispanic or Latino: American Indian and Alaska Native alone
B03002_006	Total: Not Hispanic or Latino: Asian alone
B03002_007	Total: Not Hispanic or Latino: Native Hawaiian and Other Pacific Islander alone
B03002_008	Total: Not Hispanic or Latino: Some other race alone
B03002_009	Total: Not Hispanic or Latino: Two or more races
B03002_010	Total: Not Hispanic or Latino: Two or more races: Two races including Some other race
B03002_011	Total: Not Hispanic or Latino: Two or more races: Two races excluding Some other race, and three or more races
B03002_012	Total: Hispanic or Latino
B03002_013	Total: Hispanic or Latino: White alone
B03002_014	Total: Hispanic or Latino: Black or African American alone
B03002_015	Total: Hispanic or Latino: American Indian and Alaska Native alone
B03002_016	Total: Hispanic or Latino: Asian alone
B03002_017	Total: Hispanic or Latino: Native Hawaiian and Other Pacific Islander alone
B03002_018	Total: Hispanic or Latino: Some other race alone
B03002_019	Total: Hispanic or Latino: Two or more races
B03002_020	Total: Hispanic or Latino: Two or more races: Two races including Some other race
B03002_021	Total: Hispanic or Latino: Two or more races: Two races excluding Some other race, and three or more races
}
