% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-helpers.R
\name{acs.rentburden.recode}
\alias{acs.rentburden.recode}
\title{acs.rentburden.recode}
\usage{
acs.rentburden.recode(rentb, filter.aggregates = T)
}
\arguments{
\item{rentb}{a vacancy table, as from \code{multiyr.acs.wrapper(B25004, ...)}.}

\item{filter.aggregates}{If true, filter totals and subtotals.}
}
\description{
Recodes rent burden info, as from table B25070 ("GROSS RENT AS A PERCENTAGE
OF HOUSEHOLD INCOME IN THE PAST 12 MONTHS"). Was developed for 2019 ACS;
could break if they change encodings.
}
\details{
Note: from socialexplorer.com: "Units for which no rent is paid and units
occupied by households that reported no income or a net loss comprise the
category, 'Not computed.'"

Universe for this table is Renter-occupied Housing Units.

B25070_001	Total B25070_002	Total: Less than 10.0 percent B25070_003	Total:
10.0 to 14.9 percent B25070_004	Total: 15.0 to 19.9 percent B25070_005	Total:
20.0 to 24.9 percent B25070_006	Total: 25.0 to 29.9 percent B25070_007	Total:
30.0 to 34.9 percent B25070_008	Total: 35.0 to 39.9 percent B25070_009	Total:
40.0 to 49.9 percent B25070_010	Total: 50.0 percent or more B25070_011	Total:
Not computed
}
