% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-helpers.R
\name{acs.vacancy.recode}
\alias{acs.vacancy.recode}
\title{acs.vacancy.recode}
\usage{
acs.vacancy.recode(vacancy, filter.aggregates = T)
}
\arguments{
\item{vacancy}{a vacancy table, as from \code{multiyr.acs.wrapper(B25004, ...)}.}

\item{filter.aggregates}{If true, filter totals and subtotals.}
}
\description{
Recodes demographic info, as from table B25004 Was developed for 2019 ACS;
could break if they change encodings.
}
\details{
Universe for this table is Vacant Housing Units.

B25004_001	Total
B25004_002	Total: For rent
B25004_003	Total: Rented, not occupied
B25004_004	Total: For sale only
B25004_005	Total: Sold, not occupied
B25004_006	Total: For seasonal, recreational, or occasional use
B25004_007	Total: For migrant workers
}
