% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplates.R
\name{boilerplate.git.ignore}
\alias{boilerplate.git.ignore}
\title{boilerplate.git.ignore}
\usage{
boilerplate.git.ignore()
}
\description{
Just prints a boilerplate .gitignore. Will ignore a variety of files that I end up
saving with my R projects that shouldn't be sent to github, like pictures,
documents, .zips, slurm output, saved workspaces and R objects outside of data/,
shiny artifacts, etc.
}
