% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-fcns.R
\name{count.across}
\alias{count.across}
\title{count.across}
\usage{
\method{count}{across}(x, .cols = everything())
}
\arguments{
\item{x}{data.frame}

\item{.cols}{tidy selector for given columns. By default, selects all of them.}
}
\description{
Applies the \code{count} fcn across all or a selection of columns in a data.frame.
}
