% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{gett.census.totals}
\alias{gett.census.totals}
\title{gett.census.totals}
\usage{
gett.census.totals(states, years, geo = "tract", survey = "acs5")
}
\arguments{
\item{states}{vector of state FIPS or 2-letters}

\item{years}{vector of years}

\item{geo}{selection geography, passed onto \code{tidycensus}}
}
\description{
Gets totals for area. Like population, households, housing units, and land
area. Wraps \code{tidycensus::get_acs}. 1 year at a time please.
}
