% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{mass.acs.pull.wrapper}
\alias{mass.acs.pull.wrapper}
\title{mass.acs.pull.wrapper}
\usage{
mass.acs.pull.wrapper(states, years, geo)
}
\arguments{
\item{states}{vector of state FIPS or 2-letters}

\item{years}{vector of years}

\item{geo}{selection geography, passed onto \code{tidycensus}}
}
\description{
Wrapps all the other ACS/tidycensus wrapper functions here to pull a
selection of tables with recoded values, medians, and totals, over a
selection of states and years.
}
\details{
You can use \code{taux::acs.tbl.index} for reminders of what each table is.
}
