% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-helpers.R
\name{multiyr.acs.wrapper}
\alias{multiyr.acs.wrapper}
\title{multiyr.acs.wrapper}
\usage{
multiyr.acs.wrapper(
  tables,
  states,
  geo,
  years = c(2010, 2015, 2019),
  metadata = NULL,
  cache = T,
  survey = "acs5"
)
}
\arguments{
\item{states, geo, years}{passed to \code{tidycensus::get_acs}. There is one get_acs
call for every combination of table and years.}

\item{metadata}{To use for labels. Result of \code{pull.acs.metadata} or
\code{tidycensus::load_variables}}

\item{table}{table to get from ACS; i.e., 'B01001.' See
\code{tidycensus::load_variables}}
}
\description{
Wraps \code{tidycensus::get_acs} for multiple years, tables, and states.
Structures long by geography, table, year. Maps through tables/years.
Tidycensus vectorized over states.
}
\details{
I'll need to figure out and maybe refine this function to work with
reasonable combinations of states/geographies.
}
