% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling-fcns.R
\name{named_map_dfr}
\alias{named_map_dfr}
\title{named_map_dfr}
\usage{
named_map_dfr(.x, .f, ...)
}
\description{
Uses names from a named list in an identifier column alongside other output in
map_dfr
}
