% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation-fcns.R
\name{normalize.to.range}
\alias{normalize.to.range}
\title{normalize.to.range}
\usage{
normalize.to.range(x, range = c(0, 1))
}
\description{
Normalizes a numeric vector \code{x} to between a given range
}
