% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-helpers.R
\name{pull.acs.metadata}
\alias{pull.acs.metadata}
\title{pull.acs.metadata}
\usage{
\method{pull}{acs.metadata}(year, dataset = "acs5", cache = T)
}
\description{
Pulls ACS metadata and cleans labels.
}
