% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-fcns.R
\name{quantiles.across.groups}
\alias{quantiles.across.groups}
\title{quantiles.across.groups}
\usage{
quantiles.across.groups(gx, .colm, ntile = 0.25)
}
\arguments{
\item{gx}{grouped \code{data.frame} or class that inherits}

\item{.colm}{column to get quantiles for, as string}
}
\description{
Gets (wide) quantiles by group, so each group remains one row.
}
\details{
Could improve by adding tdiy-style quotation:
https://adv-r.hadley.nz/quasiquotation.html
}
