% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling-fcns.R
\name{read.tbList}
\alias{read.tbList}
\title{read.tbList}
\usage{
read.tbList(ddir, pattern = "csv$", delim = ",")
}
\arguments{
\item{ddir}{directory containing tables to read.}

\item{pattern, delim}{passed onto \code{list.files} and \code{vroom::vroom}, respectively}
}
\description{
Read and name a tablelist, given a directory containing multiple csvs
}
