% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling-fcns.R
\name{split_and_key}
\alias{split_and_key}
\title{split_and_key}
\usage{
split_and_key(x, .cols)
}
\arguments{
\item{x}{object that inherits \code{data.frame}}

\item{cols}{a character vector of columns to split/groupby. Concatenation of these
column names, separated by "-" will form the name for each resulting df after split.}
}
\description{
Like \link{group_split} but uses grouping cols to form identifiers and uses these as
names for elements of split df. A broader purpose fcn that I may move elsewhere at
some point. Some surprise/discussion that this functionality isn't already built
into dplyr:
https://stackoverflow.com/questions/57107721/how-to-name-the-list-of-the-group-split-output-in-dplyr
}
