% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{tidycensus2recoded.tblList}
\alias{tidycensus2recoded.tblList}
\title{tidycensus2recoded.tblList}
\usage{
tidycensus2recoded.tblList(
  states,
  years,
  geo = "tract",
  tbls = c("B03002", "B25034", "B25004", "B25070"),
  metadata = NULL,
  survey = "acs5"
)
}
\arguments{
\item{states}{vector of state FIPS or 2-letters}

\item{years}{vector of years}

\item{geo}{selection geography, passed onto \code{tidycensus}}
}
\description{
Wrapper function to pull all categorical information for which there's recode
functions written. Right now this includes:
\itemize{
\item B03002 (demographics; universe is Total Pop)
\item B25034 (building age; universe is Housing Units)
\item B25004 (vacant unit types; universe is Vacant Housing Units)
\item B25070 (Rent burden counts; universe is Renter-occupied Housing Units)
\item B08006 (Commute mode; universe is Workers 16 Years and Over)
}
}
