% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling-fcns.R
\name{colm.in.tblList}
\alias{colm.in.tblList}
\title{colm.in.tblList}
\usage{
colm.in.tblList(xL, regex, return.tbls = T)
}
\arguments{
\item{xL}{list of data.frames}

\item{regex}{regular expression}

\item{return.tbls}{whether to return tables (default) or index vector}
}
\description{
Given a list of data.frames, find which ones have a column that match a regex string.
}
