% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{pull.tidycensus.median.tables}
\alias{pull.tidycensus.median.tables}
\title{pull.tidycensus.median.tables}
\usage{
\method{pull}{tidycensus.median.tables}(
  states,
  years,
  geo = "tract",
  tbls = c("B19013", "B25058"),
  survey = "acs5"
)
}
\arguments{
\item{states}{vector of state FIPS or 2-letters}

\item{years}{vector of years}

\item{geo}{selection geography, passed onto \code{tidycensus}}
}
\description{
Pulls median tables, like hh income and contract rent.
}
